requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
httpx==0.27.0
python-dotenv==1.0.1
python-multipart==0.0.9
